Breast Cancer Identification/requirement.txt
pandas
numpy
matplotlib
seaborn
Pillow
tensorflow
keras
scikit-learn
h5py
requests
streamlit